requirements.txt
beartype==0.21.0
einops==0.8.1
matplotlib==3.7.2
numpy==2.3.1
opencv_python==4.8.1.78
Pillow==10.0.1
Pillow==11.2.1
scipy==1.11.3
shortuuid==1.0.13
tensorboardX==2.6.2.2
tensorboardX==2.6.4
timm==0.9.7
torch==2.1.0
torchsummary==1.5.1
torchvision==0.16.0
tqdm==4.66.4
tqdm==4.67.0
